affection/requirements.txt
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.5.1
scipy==1.11.4
imbalanced-learn==0.12.0
joblib==1.3.2
matplotlib
seaborn

# Azure + Cosmos DB
azure-core==1.29.5
azure-cosmos==4.5.1
azure-identity==1.16.0
azureml-defaults

# For JSON logs & utilities
pyyaml
pyodbc
